added base viewmodel, base usecase, changed save user functionality, refactored some names.
