+++ app/src/main/java/com/space/quiz/QuizApp.kt
@@ -1,10 +1,7 @@
 package com.space.quiz
 
 import android.app.Application
-import com.space.quiz.di.dbModule
-import com.space.quiz.di.repositoryModule
-import com.space.quiz.di.useCaseModule
-import com.space.quiz.di.viewModelModule
+import com.space.quiz.di.*
 import org.koin.android.ext.koin.androidContext
 import org.koin.android.ext.koin.androidLogger
 import org.koin.core.context.startKoin
@@ -16,7 +13,7 @@ class QuizApp : Application() {
             androidLogger()
             androidContext(this@QuizApp)
             modules(
-                dbModule, repositoryModule,useCaseModule,viewModelModule
+                dbModule, repositoryModule,useCaseModule,viewModelModule, mapperModule
             )
         }
     }
