+++ app/src/main/java/com/space/quiz/data/local/mapper/UserDomainToEntityMapper.kt
@@ -1,13 +1,16 @@
 package com.space.quiz.data.local.mapper
 
 import com.space.quiz.data.local.entity.UserEntity
-import com.space.quiz.domain.model.DomainUserModel
-import com.space.quiz.utils.ModelMapper
+import com.space.quiz.domain.model.UserDomainModel
+import com.space.quiz.utils.UIMapper
 
-class UserDomainToEntityMapper :ModelMapper<DomainUserModel,UserEntity>{
-    override fun invoke(model: DomainUserModel): UserEntity =
-        UserEntity(
-            id = model.id,
-            userName = model.userName
-        )
+class UserDomainToEntityMapper :UIMapper<UserDomainModel,UserEntity>{
+    override fun invoke(model: UserDomainModel): UserEntity =
+        with(model){
+            UserEntity(
+                id = id,
+                userName = userName
+            )
+        }
+
 }
