+++ app/src/main/java/com/space/quiz/data/local/mapper/UserEntityToDomainMapper.kt
@@ -1,13 +1,16 @@
 package com.space.quiz.data.local.mapper
 
 import com.space.quiz.data.local.entity.UserEntity
-import com.space.quiz.domain.model.DomainUserModel
-import com.space.quiz.utils.ModelMapper
+import com.space.quiz.domain.model.UserDomainModel
+import com.space.quiz.utils.UIMapper
 
-class UserEntityToDomainMapper : ModelMapper<UserEntity,DomainUserModel>{
-    override fun invoke(model: UserEntity): DomainUserModel =
-        DomainUserModel(
-            id = model.id,
-            userName = model.userName
-        )
+class UserEntityToDomainMapper : UIMapper<UserEntity,UserDomainModel>{
+    override fun invoke(model: UserEntity): UserDomainModel =
+        with(model){
+            UserDomainModel(
+                id = id,
+                userName = userName
+            )
+        }
+
 }
