+++ app/src/main/java/com/space/quiz/di/MapperModule.kt
@@ -1,0 +1,8 @@
+package com.space.quiz.di
+
+import com.space.quiz.presentation.model.mapper.UserUIDomainMapper
+import org.koin.dsl.module
+
+val mapperModule = module {
+    single { UserUIDomainMapper() }
+}
