+++ app/src/main/java/com/space/quiz/di/UseCaseModule.kt
@@ -1,9 +1,11 @@
 package com.space.quiz.di
 
-import com.space.quiz.domain.usecase.save_user.SaveUserUseCase
+import com.space.quiz.domain.model.UserDomainModel
+import com.space.quiz.domain.usecase.base.BaseUseCase
 import com.space.quiz.domain.usecase.save_user.SaveUserUseCaseImpl
+import com.space.quiz.utils.Resource
 import org.koin.dsl.module
 
 val useCaseModule = module {
-    single<SaveUserUseCase> { SaveUserUseCaseImpl(get()) }
+    single<BaseUseCase<UserDomainModel, Resource>> { SaveUserUseCaseImpl(get()) }
 }
