+++ app/src/main/java/com/space/quiz/di/ViewModelModule.kt
@@ -1,24 +1,26 @@
 package com.space.quiz.di
 
-import com.space.quiz.presentation.fragment_details.vm.DetailsViewModel
-import com.space.quiz.presentation.fragment_home.vm.HomeViewModel
-import com.space.quiz.presentation.fragment_intro.vm.IntroViewModel
-import com.space.quiz.presentation.fragment_questions.vm.QuestionsViewModel
-import com.space.quiz.presentation.model.mapper.UserPresentationToDomainMapper
+import com.space.quiz.presentation.ui_details.vm.DetailsViewModel
+import com.space.quiz.presentation.ui_home.vm.HomeViewModel
+import com.space.quiz.presentation.ui_intro.vm.IntroViewModel
+import com.space.quiz.presentation.ui_questions.vm.QuestionsViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
 
 val viewModelModule = module {
-    viewModel{
-        IntroViewModel(get(), UserPresentationToDomainMapper())
+    viewModel {
+        IntroViewModel(
+            get(),
+            get()
+        )
     }
-    viewModel{
+    viewModel {
         HomeViewModel()
     }
-    viewModel{
+    viewModel {
         DetailsViewModel()
     }
-    viewModel{
+    viewModel {
         QuestionsViewModel()
     }
 }
