--- app/src/main/java/com/space/quiz/domain/model/DomainUserModel.kt
@@ -1,6 +1,0 @@
-package com.space.quiz.domain.model
-
-data class DomainUserModel(
-    val id: Int? = null,
-    val userName: String
-)
