+++ app/src/main/java/com/space/quiz/domain/model/UserDomainModel.kt
@@ -1,0 +1,6 @@
+package com.space.quiz.domain.model
+
+data class UserDomainModel(
+    val id: Int? = null,
+    val userName: String
+)
