+++ app/src/main/java/com/space/quiz/domain/repository/UserRepository.kt
@@ -1,8 +1,8 @@
 package com.space.quiz.domain.repository
 
-import com.space.quiz.domain.model.DomainUserModel
+import com.space.quiz.domain.model.UserDomainModel
 
 interface UserRepository {
-    suspend fun saveUser(user: DomainUserModel)
-    suspend fun getUserName(userName:String): DomainUserModel?
+    suspend fun saveUser(user: UserDomainModel)
+    suspend fun getUserName(userName:String): UserDomainModel?
 }
