+++ app/src/main/java/com/space/quiz/domain/usecase/base/BaseUseCase.kt
@@ -1,0 +1,5 @@
+package com.space.quiz.domain.usecase.base
+
+abstract class BaseUseCase<in Params, out T> {
+    abstract suspend operator fun invoke(params: Params? = null): T
+}
