--- app/src/main/java/com/space/quiz/presentation/fragment_details/vm/DetailsViewModel.kt
@@ -1,6 +1,0 @@
-package com.space.quiz.presentation.fragment_details.vm
-
-import androidx.lifecycle.ViewModel
-
-class DetailsViewModel():ViewModel() {
-}
