--- app/src/main/java/com/space/quiz/presentation/fragment_home/vm/HomeViewModel.kt
@@ -1,6 +1,0 @@
-package com.space.quiz.presentation.fragment_home.vm
-
-import androidx.lifecycle.ViewModel
-
-class HomeViewModel():ViewModel() {
-}
