--- app/src/main/java/com/space/quiz/presentation/fragment_questions/vm/QuestionsViewModel.kt
@@ -1,6 +1,0 @@
-package com.space.quiz.presentation.fragment_questions.vm
-
-import androidx.lifecycle.ViewModel
-
-class QuestionsViewModel():ViewModel() {
-}
