--- app/src/main/java/com/space/quiz/presentation/model/PresentationUserModel.kt
@@ -1,6 +1,0 @@
-package com.space.quiz.presentation.model
-
-data class PresentationUserModel(
-    val id: Int? = null,
-    val userName: String
-)
