+++ app/src/main/java/com/space/quiz/presentation/model/UserUIModel.kt
@@ -1,0 +1,6 @@
+package com.space.quiz.presentation.model
+
+data class UserUIModel(
+    val id: Int? = null,
+    val userName: String
+)
