+++ app/src/main/java/com/space/quiz/presentation/ui_details/vm/DetailsViewModel.kt
@@ -1,0 +1,7 @@
+package com.space.quiz.presentation.ui_details.vm
+
+import androidx.lifecycle.ViewModel
+import com.space.quiz.presentation.base.BaseViewModel
+
+class DetailsViewModel():BaseViewModel() {
+}
