+++ app/src/main/java/com/space/quiz/presentation/ui_intro/vm/IntroViewModel.kt
@@ -1,0 +1,33 @@
+package com.space.quiz.presentation.ui_intro.vm
+
+import com.space.quiz.domain.model.UserDomainModel
+import com.space.quiz.domain.usecase.base.BaseUseCase
+import com.space.quiz.presentation.base.BaseViewModel
+import com.space.quiz.presentation.model.UserUIModel
+import com.space.quiz.presentation.model.mapper.UserUIDomainMapper
+import com.space.quiz.presentation.ui_intro.ui.IntroFragmentDirections
+import com.space.quiz.utils.Resource
+import com.space.quiz.utils.viewModelScope
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
+
+class IntroViewModel(
+    private val saveUserUseCase: BaseUseCase<UserDomainModel, Resource>,
+    private val userUIToDomain: UserUIDomainMapper
+) : BaseViewModel() {
+    private val _errorFlow = MutableStateFlow(0)
+    val errorFlow: StateFlow<Int> = _errorFlow
+
+    fun authenticateUser(user: UserUIModel) {
+        viewModelScope {
+            when (val status = saveUserUseCase.invoke(userUIToDomain(user))) {
+                is Resource.Success -> {
+                    navigate(IntroFragmentDirections.actionIntroFragmentToHomeFragment())
+                }
+                is Resource.Error -> {
+                    _errorFlow.value = (status.errorMessage)
+                }
+            }
+        }
+    }
+}
