--- app/src/main/java/com/space/quiz/utils/ModelMapper.kt
@@ -1,5 +1,0 @@
-package com.space.quiz.utils
-
-interface ModelMapper<in MODEL_A, out MODEL_B> {
-    operator fun invoke(model: MODEL_A): MODEL_B
-}
