+++ app/src/main/java/com/space/quiz/utils/Resource.kt
@@ -3,6 +3,7 @@ package com.space.quiz.utils
 import androidx.annotation.StringRes
 
 sealed class Resource{
-    data class Error(@StringRes val errorMessage : Int):Resource()
+    data class Error(val errorMessage : Int):Resource()
     object Success : Resource()
 }
+
