+++ app/src/main/java/com/space/quiz/utils/ToastExtension.kt
@@ -1,0 +1,8 @@
+package com.space.quiz.utils
+
+import android.content.Context
+import android.widget.Toast
+
+fun Context.showToast(message: String, duration: Int = Toast.LENGTH_LONG) {
+    Toast.makeText(this, message, duration).show()
+}
