+++ app/src/main/java/com/space/quiz/utils/UIMapper.kt
@@ -1,0 +1,5 @@
+package com.space.quiz.utils
+
+interface UIMapper<in MODEL_A, out MODEL_B> {
+    operator fun invoke(model: MODEL_A): MODEL_B
+}
