+++ app/src/main/res/layout/custom_dialog.xml
@@ -77,7 +77,7 @@
         android:layout_marginTop="@dimen/dimen_14dp"
         android:layout_marginBottom="@dimen/dimen_27dp"
         android:padding="@dimen/dimen_3dp"
-        android:text="@string/text_close"
+        android:text="@string/close_text"
         app:layout_constraintBottom_toBottomOf="@+id/dialogRectangleView"
         app:layout_constraintEnd_toEndOf="@+id/viewLine"
         app:layout_constraintStart_toStartOf="@+id/viewLine" />
