+++ app/src/main/res/layout/fragment_home.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.fragment_home.ui.HomeFragment">
+    tools:context=".presentation.ui_home.ui.HomeFragment">
 
     <androidx.appcompat.widget.AppCompatTextView
         android:id="@+id/helloUserTextView"
@@ -43,7 +43,7 @@
         style="@style/largeMyriadGeo_white_normal"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="@string/gpa_4_0"
+        android:text="@string/gpa_text"
         app:layout_constraintBottom_toBottomOf="@+id/gpaRectangle"
         app:layout_constraintEnd_toEndOf="@+id/gpaRectangle"
         app:layout_constraintStart_toStartOf="@+id/gpaRectangle"
